Version checking
When loading a file, program now checks the versions (program version and file version) and compares it. When both versions are the same, file is loaded. When not, program checks the list of older versions. When the file version is on the list, program checks changes in the file data system to convert file (currently not working due to constant data system). When the file version was not found on the list, program can not open the file and processing is aborted.
Also fixed some typos and changed some variables to constant.